[MTGOSDK/Core, API] Refactor ObjectProvider type binding, User API reflection caching
This refactor binds a class's interface to the returned remote object from ObjectProvider.

This however introduces issues with reflection caching, as any access to MemberInfo derived properties (i.e. FullName) will cache the local type's field properties. This conflicts with the RemoteType constructor's reflection.

This User API refactor works around calling overloaded methods (not supported) and reflection caching issues as it needs to be called when constructing the Client class.