[MTGOSDK/API] ObjectProvider: Enable lazy instantiation
Defaults to lazy instantiation of dynamic remote objects, deferring instantiation to a later call to ObjectProvider.

This implements the same functionality between RemoteProxy and LazyRemoteObject, allowing for multiple lazy initialization schemes to be daisy chained.

This cover cases where an object's members would need to be enumerated to bind an interface type, and also where only the object reference is needed.