[MTGOSDK.Tests] SetupFixture: Always skip online check
Remove census API check before starting MTGO (via `ClientOptions.CreateProcess`).

The `Client.IsOnline()` method may be unreliable due to discrepancies in reporting w/ the census API. This method is still useful when combined with other sanity checks (which are use-case specific), we'll only remove the check and leave the `IsOnline()` method for now.