[MTGOSDK/Core/Compiler] ObjectFreezer: Allow pinning private types
Modifies the IL code generation to avoid explicit type casting, instead pinning the object as a reference to `object`.

We instead declaring declare pinned local variable as `object.MakeByRefType()` instead of the specific type and directly storing the input object into it. This allows the pinning of the object's reference, bypassing the type accessibility checks performed by `OpCodes.Castclass`.