Disable snupkg creation
As we already bundle PDBs in the main nupkg and have configured SourceLink to point to this repository, distributing snupkg files is mostly redundant.

As we do not enable SourceLink for ScubaDiver, disabling this effectively supersedes #15 and resolves #32 and no longer relies on nuget.org specific behavior regarding symbol fetching.